+++ docs/immersive-translate.user.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         Immersive Translate
 // @description  Web bilingual translation, completely free to use, supports Deepl/Google/Bing/Tencent/Youdao, etc. it also works on iOS Safari.
-// @version      0.3.15
+// @version      0.4.0
 // @namespace    https://immersive-translate.owenyoung.com/
 // @author       Owen Young
 // @homepageURL    https://immersive-translate.owenyoung.com/
@@ -74,7 +74,7 @@
   };
 
   // <define:process.env>
-  var define_process_env_default = { BUILD_TIME: "2023-03-31T22:28:33.204Z", VERSION: "0.3.15", PROD: "1", REDIRECT_URL: "https://immersive-translate.owenyoung.com/auth-done/", IMMERSIVE_TRANSLATE_INJECTED_CSS: `:root {
+  var define_process_env_default = { BUILD_TIME: "2023-04-06T09:54:37.467Z", VERSION: "0.4.0", PROD: "1", REDIRECT_URL: "https://immersive-translate.owenyoung.com/auth-done/", IMMERSIVE_TRANSLATE_INJECTED_CSS: `:root {
   --immersive-translate-theme-underline-borderColor: #72ece9;
   --immersive-translate-theme-nativeUnderline-borderColor: #72ece9;
   --immersive-translate-theme-nativeDashed-borderColor: #72ece9;
@@ -3839,7 +3839,7 @@ body {
   </button>
   <div class="immersive-translate-popup-mount" id="mount"></div>
 </div>
-`, OPTIONS_URL: "https://immersive-translate.owenyoung.com/options/", EBOOK_VIEWER_URL: "https://immersive-translate.owenyoung.com/ebook/", EBOOK_BUILDER_URL: "https://immersive-translate.owenyoung.com/ebook/make/", HTML_VIEWER_URL: "https://immersive-translate.owenyoung.com/html/", MOCK: "0", DEBUG: "0", IMMERSIVE_TRANSLATE_USERSCRIPT: "1" };
+`, OPTIONS_URL: "https://immersive-translate.owenyoung.com/options/", EBOOK_VIEWER_URL: "https://immersive-translate.owenyoung.com/ebook/", EBOOK_BUILDER_URL: "https://immersive-translate.owenyoung.com/ebook/make/", SUBTITLE_BUILDER_URL: "https://immersive-translate.owenyoung.com/subtitle/", HTML_VIEWER_URL: "https://immersive-translate.owenyoung.com/html/", MOCK: "0", DEBUG: "0", IMMERSIVE_TRANSLATE_USERSCRIPT: "1" };
 
   // https://esm.sh/v113/n-gram@2.0.2/deno/n-gram.mjs
   var c = o(2), f = o(3);
@@ -4229,6 +4229,13 @@ body {
     else
       throw new Error("ebook url not found");
   }
+  function openSubtitleBuilderPage(newTab = !1) {
+    let optionsUrl = getEnv().SUBTITLE_BUILDER_URL;
+    if (optionsUrl)
+      openInTab(optionsUrl, newTab);
+    else
+      throw new Error("subtitle url not found");
+  }
   function getAcceptLanguages() {
     return globalThis.navigator.languages || // @ts-ignore: it's ok
     [globalThis.navigator.language || globalThis.navigator.userLanguage];
@@ -4247,7 +4254,8 @@ body {
       detectLanguage: languageDetect,
       openAboutPage,
       openEbookViewerPage,
-      openEbookBuilderPage
+      openEbookBuilderPage,
+      openSubtitleBuilderPage
     },
     i18n: {
       getAcceptLanguages,
@@ -4418,8 +4426,9 @@ body {
     "browser.translateLocalPdfFile": "\u7FFB\u8BD1\u672C\u5730 PDF \u6587\u4EF6",
     "browser.openEbookViewer": "\u9605\u8BFB\u672C\u5730\u7535\u5B50\u4E66",
     "browser.openEbookBuilder": "\u5236\u4F5C\u53CC\u8BED Epub \u7535\u5B50\u4E66",
-    "browser.translateLocalHtmlFile": "\u7FFB\u8BD1\u672C\u5730 HTML \u6587\u4EF6",
+    "browser.translateLocalHtmlFile": "\u7FFB\u8BD1 HTML/txt \u6587\u4EF6",
     "browser.donateContext": "\u4E86\u89E3\u8D5E\u52A9\u798F\u5229",
+    "browser.translateLocalSubtitleFile": "\u7FFB\u8BD1\u672C\u5730\u5B57\u5E55\u6587\u4EF6",
     confirmResetConfig: "\u4F60\u786E\u5B9A\u8981\u91CD\u7F6E\u8BBE\u7F6E\u5417\uFF1F",
     translationLineBreakSettingTitle: "\u8BD1\u6587\u6362\u884C\u8BBE\u7F6E",
     smartLineBreak: "\u667A\u80FD\u6362\u884C",
@@ -4697,7 +4706,8 @@ body {
     "browser.translateLocalPdfFile": "\u7FFB\u8B6F\u672C\u6A5F PDF \u6A94\u6848",
     "browser.openEbookViewer": "\u95B1\u8B80\u672C\u6A5F\u96FB\u5B50\u66F8",
     "browser.openEbookBuilder": "\u88FD\u4F5C\u96D9\u8A9E Epub \u96FB\u5B50\u66F8",
-    "browser.translateLocalHtmlFile": "\u7FFB\u8B6F\u672C\u6A5F HTML \u6A94\u6848",
+    "browser.translateLocalHtmlFile": "\u7FFB\u8B6F HTML/txt \u6A94\u6848",
+    "browser.translateLocalSubtitleFile": "\u7FFB\u8B6F\u5B57\u5E55\u6A94\u6848",
     "browser.donateContext": "\u4E86\u89E3\u8D0A\u52A9\u798F\u5229",
     confirmResetConfig: "\u4F60\u78BA\u5B9A\u8981\u91CD\u8A2D\u8A2D\u5B9A\u55CE\uFF1F",
     translationLineBreakSettingTitle: "\u8B6F\u6587\u63DB\u884C\u8A2D\u5B9A",
@@ -4967,11 +4977,12 @@ body {
     "browser.toggleTranslateTheMainPage": "Toggle translate the main page",
     "browser.openOptionsPage": "Open Settings Page",
     "browser.toggleTranslationMask": "Toggle translation mask style",
-    "browser.translateLocalPdfFile": "Translate local PDF files",
+    "browser.translateLocalPdfFile": "Translate local PDF File",
     "browser.openEbookViewer": "Read local e-book",
     "browser.openEbookBuilder": "Make Dual Epub ebook",
-    "browser.translateLocalHtmlFile": "Translate local HTML files",
+    "browser.translateLocalHtmlFile": "Translate HTML/txt File",
     "browser.donateContext": "Sponsor Benefits",
+    "browser.translateLocalSubtitleFile": "Translate Subtitle File",
     confirmResetConfig: "Are you sure you want to reset the settings?",
     translationLineBreakSettingTitle: "Line break setting",
     smartLineBreak: "Smart Wrap",
@@ -5243,7 +5254,8 @@ body {
   var brandIdForJs = "immersiveTranslate";
   var GOOGLE_ACCESS_TOKEN_KEY = brandIdForJs + "GoogleAccessToken", AUTH_FLOW_FLAG = brandIdForJs + "AuthFlow";
   var AUTH_STATE_FLAG = brandIdForJs + "AuthState", iframeMessageIdentifier = brandIdForJs + "IframeMessage", iframeMessageRateIdentifier = brandIdForJs + "WaitForRateLimit", documentMessageTypeIdentifierForAsk = brandIdForJs + "DocumentMessageAsk", documentMessageTypeIdentifierForTellThirdParty = brandIdForJs + "DocumentMessageTellThirdParty", documentMessageTypeIdentifierForThirdPartyTell = brandIdForJs + "DocumentMessageThirdPartyTell", documentMessageTypeIdentifierForHandler = brandIdForJs + "DocumentMessageHandler", targetContainerElementAttributeName = `${brandIdForJs}Container`, specifiedTargetContainerElementAttributeName = `${brandIdForJs}SpecifiedContainer`, buildinConfigStorageKey = "buildinConfig", localConfigStorageKey = "localConfig", contextOpenOptionsMenuId = "openOptionsPage", contextOpenAboutMenuId = "openAboutPage";
-  var contextOpenLocalEbookViewer = "openEbookViewer", contextOpenLocalEbookBuilder = "openEbookBuilder", pageTranslatedStatusEventName = `${brandIdForJs}PageTranslatedStatus`, pageUrlChangedEventName = `${brandIdForJs}PageUrlChanged`, userscriptCommandEventName = `${brandIdForJs}ReceiveCommand`, popupReceiveMessageEventName = `${brandIdForJs}PopupReceiveMessage`, hostname = "immersive-translate.owenyoung.com", homepage = `https://${hostname}/`, buildinConfigSyncUrl = `https://${hostname}/buildin_config.json`, sourceElementMarkAttributeName = `${brandIdForJs}Mark`, sourceElementEffectAttributeNameForJs = "immersiveTranslateEffect", elementMarkRootKey = `${brandIdForJs}Root`, sourceElementEffectAttributeName = `data-${brandId}-effect`, sourceElementTranslatedMarkAttributeName = `${brandIdForJs}TranslatedMark`, sourceElementParagraphAttributeName = `${brandIdForJs}ParagraphId`, sourceAtomicBlockElementMarkAttributeName = `${brandIdForJs}AtomicBlockMark`, sourceElementExcludeAttributeName = `${brandIdForJs}ExcludeMark`, sourceElementExcludeAttributeNameForSelector = `data-${brandId}-exclude-mark`, sourceElementStayOriginalAttributeName = `${brandIdForJs}StayOriginalMark`, sourcePreWhitespaceMarkAttributeName = `${brandIdForJs}PreWhitespaceMark`, sourceInlineElementMarkAttributeName = `${brandIdForJs}InlineMark`, sourceBlockElementMarkAttributeName = `${brandIdForJs}BlockMark`, sourceElementLeft = `${brandIdForJs}Left`, sourceElementRight = `${brandIdForJs}Right`, sourceElementWidth = `${brandIdForJs}Width`, sourceElementHeight = `${brandIdForJs}Height`, sourceElementTop = `${brandIdForJs}Top`, sourceElementFontSize = `${brandIdForJs}FontSize`, lastRunTimeStorageKey = "lastRunTime", sourceElementWithGlobalStyleMarkAttributeName = `${brandIdForJs}GlobalStyleMark`, defaultPlaceholderDelimiters = ["@", "#"], titleDelimiters = " --- ", translationTextSeparator = `
+  var contextOpenLocalEbookViewer = "openEbookViewer", contextOpenLocalEbookBuilder = "openEbookBuilder";
+  var pageTranslatedStatusEventName = `${brandIdForJs}PageTranslatedStatus`, pageUrlChangedEventName = `${brandIdForJs}PageUrlChanged`, userscriptCommandEventName = `${brandIdForJs}ReceiveCommand`, popupReceiveMessageEventName = `${brandIdForJs}PopupReceiveMessage`, hostname = "immersive-translate.owenyoung.com", homepage = `https://${hostname}/`, buildinConfigSyncUrl = `https://${hostname}/buildin_config.json`, sourceElementMarkAttributeName = `${brandIdForJs}Mark`, sourceElementEffectAttributeNameForJs = "immersiveTranslateEffect", elementMarkRootKey = `${brandIdForJs}Root`, sourceElementEffectAttributeName = `data-${brandId}-effect`, sourceElementTranslatedMarkAttributeName = `${brandIdForJs}TranslatedMark`, sourceElementParagraphAttributeName = `${brandIdForJs}ParagraphId`, sourceAtomicBlockElementMarkAttributeName = `${brandIdForJs}AtomicBlockMark`, sourceElementExcludeAttributeName = `${brandIdForJs}ExcludeMark`, sourceElementExcludeAttributeNameForSelector = `data-${brandId}-exclude-mark`, sourceElementStayOriginalAttributeName = `${brandIdForJs}StayOriginalMark`, sourcePreWhitespaceMarkAttributeName = `${brandIdForJs}PreWhitespaceMark`, sourceInlineElementMarkAttributeName = `${brandIdForJs}InlineMark`, sourceBlockElementMarkAttributeName = `${brandIdForJs}BlockMark`, sourceElementLeft = `${brandIdForJs}Left`, sourceElementRight = `${brandIdForJs}Right`, sourceElementWidth = `${brandIdForJs}Width`, sourceElementHeight = `${brandIdForJs}Height`, sourceElementTop = `${brandIdForJs}Top`, sourceElementFontSize = `${brandIdForJs}FontSize`, lastRunTimeStorageKey = "lastRunTime", sourceElementWithGlobalStyleMarkAttributeName = `${brandIdForJs}GlobalStyleMark`, defaultPlaceholderDelimiters = ["@", "#"], titleDelimiters = " --- ", translationTextSeparator = `
 `, translationTargetElementWrapperClass = `${brandId}-target-wrapper`, translationPdfTargetContainerClass = `${brandId}-pdf-target-container`, translationTargetInnerElementWrapperClass = `${brandId}-target-inner`, translationSourceElementsWrapperClass = `${brandId}-source-wrapper`, translationTargetTranslationElementBlockWrapperClass = `${brandId}-target-translation-block-wrapper`, translationFrameRootThemeAttributeName = `${brandId}-root-translation-theme`, translationFrameRootThemeAttributeNameForJs = `${brandIdForJs}RootTranslationTheme`, translationTargetTranslationElementVerticalBlockClass = `${brandId}-target-translation-vertical-block-wrapper`, translationTargetTranslationPdfElementBlockWrapperClass = `${brandId}-target-translation-pdf-block-wrapper`, translationTargetTranslationElementPreWhitespaceWrapperClass = `${brandId}-target-translation-pre-whitespace`, translationTargetTranslationElementInlineWrapperClass = `${brandId}-target-translation-inline-wrapper`;
   var themeOptions = {
     underline: [
@@ -5815,7 +5827,8 @@ body {
     },
     papago: {
       name: "Papago",
-      homepage: "https://translate.google.com/"
+      homepage: "https://translate.google.com/",
+      canary: !0
     },
     baidu: {
       name: "Baidu",
@@ -8034,7 +8047,18 @@ body {
     return null;
   }
   function getWhitespace(nextNode, isPreWhitespace) {
-    return nextNode && nextNode.nodeType === Node.TEXT_NODE && nextNode.textContent && nextNode.textContent?.length > 0 ? isPreWhitespace ? nextNode.textContent : " " : nextNode && nextNode.nodeType === Node.ELEMENT_NODE ? " " : null;
+    return nextNode && nextNode.nodeType === Node.TEXT_NODE && nextNode.textContent && nextNode.textContent?.length > 0 ? isPreWhitespace ? nextNode.textContent : " " : null;
+  }
+  function customTrim(str) {
+    if (!str)
+      return "";
+    let nbspChar = String.fromCharCode(160);
+    if (str.includes(nbspChar)) {
+      let trimedStr = str.trim(), leftPosition = str.indexOf(trimedStr), rightPosition = leftPosition + trimedStr.length, leftStr = str.substring(0, leftPosition), rightStr = str.substring(rightPosition), leftNbspCount = leftStr.split(nbspChar).length - 1, rightNbspCount = rightStr.split(nbspChar).length - 1, finalLeftStr = leftNbspCount > 0 ? " ".repeat(leftNbspCount) : "", finalRightStr = rightNbspCount > 0 ? " ".repeat(rightNbspCount) : "";
+      return finalLeftStr + trimedStr + finalRightStr;
+    } else
+      str = str.trim();
+    return str;
   }
   function getElementsBySelectors(root2, selectors) {
     let elements = [];
@@ -8178,6 +8202,19 @@ body {
       let doc = iframes[i3].contentDocument;
       if (doc && (text += getMainText(doc.body), text.length > 2e3))
         break;
+    }
+    return text;
+  }
+  function getPairsSourceText(root2, pairs) {
+    let keys = Object.keys(pairs), text = "";
+    for (let i3 = 0; i3 < keys.length; i3++) {
+      let sourceSelector = keys[i3], elements = root2.querySelectorAll(sourceSelector);
+      for (let j6 = 0; j6 < elements.length; j6++) {
+        let element = elements[j6];
+        if (text += (element.innerText || element.textContent || "") + `
+`, text.length > 2e3)
+          break;
+      }
     }
     return text;
   }
@@ -8678,7 +8715,9 @@ body {
         translationDebounce: 300,
         limit: 1500,
         maxTextGroupLengthPerRequest: 1,
-        prompt: "Do not explain, translate the text below to {{to}}:",
+        prompt: `Translate the text to {{to}}:
+
+{{text}}`,
         newlinePlaceholderDelimiters: [
           `
 
@@ -8768,6 +8807,8 @@ body {
       isEbookBuilder: !1,
       waitForSelectors: [],
       waitForSelectorsTimeout: 3e3,
+      pairs: {},
+      isSubtitleBuilder: !1,
       minZIndex: 0,
       initTranslationServiceAsSoonAsPossible: !0,
       targetWrapperTag: "font",
@@ -8788,7 +8829,8 @@ body {
         ".notranslate",
         ".MathJax_Preview",
         ".MathJax",
-        ".mwe-math-element"
+        ".mwe-math-element",
+        "[translate=no]"
       ],
       translationBlockStyle: "",
       isShowUserscriptPagePopup: !0,
@@ -8854,7 +8896,8 @@ body {
         "table.highlight",
         "div[class^=codeBlockContent]",
         "div[class^=codeBlockLines]",
-        "div[class^=token-line]"
+        "div[class^=token-line]",
+        "#liuchan-window > .liuchan-container > *"
       ],
       translationClasses: [],
       atomicBlockSelectors: [],
@@ -9076,6 +9119,15 @@ body {
         useIframePostMessage: !1
       },
       {
+        matches: [
+          "googleads.g.doubleclick.net",
+          "https://www.google.com/recaptcha/*",
+          "ad.doubanio.com"
+        ],
+        useIframePostMessage: !1,
+        selectors: "#notexistforimmersivetranslate"
+      },
+      {
         matches: ["mail.jabber.org", "antirez.com"],
         excludeTags: [
           "TITLE",
@@ -9255,8 +9307,9 @@ body {
         matches: "github.com",
         observeUrlChange: !0,
         excludeMatches: [
-          "https://github.com/settings/profile",
-          "https://github.com/*/*/settings"
+          "https://github.com/*/*/settings",
+          "https://github.com/settings/*",
+          "https://github.com/sponsors/*"
         ],
         selectors: [
           ".markdown-title",
@@ -9466,10 +9519,16 @@ body {
         ]
       },
       {
-        matches: ["scholar.google.com"],
+        matches: ["scholar.google.com", "scholar.google.com.hk"],
         wrapperPrefix: `
 `,
-        selectors: ["h3 a[data-clk]", "div.gs_rs"],
+        selectors: [
+          "h3 a[data-clk]",
+          "div.gs_rs",
+          "td a.gsc_a_at",
+          "td div.gs_gray:last-of-type",
+          "div.gsc_oci_value"
+        ],
         atomicBlockSelectors: [".gs_rs", "h3 a[data-clk]"]
       },
       {
@@ -9712,7 +9771,8 @@ body {
       },
       {
         matches: "www.newyorker.com",
-        additionalSelectors: ["h1", "[data-testid=SummaryItemHed]"]
+        additionalSelectors: ["h1", "[data-testid=SummaryItemHed]"],
+        urlChangeDelay: 2e3
       },
       {
         matches: "start.me",
@@ -10400,12 +10460,6 @@ body {
         ]
       },
       {
-        matches: "doc.qt.io",
-        additionalStayOriginalSelectors: [
-          "[translate=no]"
-        ]
-      },
-      {
         matches: "wandb.ai",
         additionalSelectors: [".report-page-top"],
         additionalStayOriginalSelectors: [
@@ -10419,6 +10473,24 @@ body {
           "font[face=verdana]"
         ],
         extraBlockSelectors: "font[face=verdana]"
+      },
+      {
+        matches: "*.zendesk.com",
+        additionalSelectors: [
+          "div.zd-comment"
+        ]
+      },
+      {
+        isSubtitleBuilder: !0,
+        selectorMatches: [
+          "meta[name='immersive-translate-subtitle-builder'][content='true']"
+        ],
+        pairs: {
+          ".source-text": ".target-text"
+        },
+        isTranslateTitle: !1,
+        wrapperPrefix: "",
+        wrapperSuffix: ""
       }
     ]
   };
@@ -11486,6 +11558,11 @@ body {
       method: "openEbookBuilderPage"
     });
   }
+  function openSubtitleBuilderPage2(newTab = !1) {
+    return isMonkey() ? (browserAPI.extra.openSubtitleBuilderPage(newTab), Promise.resolve()) : sendMessage({
+      method: "openSubtitleBuilderPage"
+    });
+  }
   function autoSyncLatestConfig() {
     return isMonkey() ? (log_default.warn("autoSyncLatestConfig is not support in monkey"), Promise.resolve()) : sendMessage({
       method: "autoSyncLatestConfig"
@@ -11507,6 +11584,7 @@ body {
     let { rule, state: { translationArea } } = ctx, contentContainers = [];
     if (translationArea === "body")
       return [root2];
+    let pairsKeys = Object.keys(rule.pairs);
     if (rule && rule.selectors.length > 0) {
       let containers = rule.selectors.map((selector) => {
         let isMatch = root2.matches(selector), elements = [];
@@ -11729,7 +11807,7 @@ ${injectedCss}}
       }
       if (!isExcludeElement(element, rule, !0)) {
         {
-          let finalText = isPreWhitespace ? rawText : rawText.trim().replace(/\n/g, " ");
+          let finalText = isPreWhitespace ? rawText : customTrim(rawText).replace(/\n/g, " ");
           if (isUrl(finalText) || isHashTag(finalText) || isAtTag(finalText) || isStockTag(finalText)) {
             let variable = {
               type: "element",
@@ -12102,8 +12180,42 @@ ${injectedCss}}
     };
   }
 
+  // dom/get_pairs_paragraphs.ts
+  function getParagraphs2(rootFrame, ctx) {
+    let pariKeys = Object.keys(ctx.rule.pairs), allParagraphs = [], id = 1, currentPageLanguageByClient2 = "auto", currentPageLanguage2 = getCurrentPageLanguage();
+    ctx.state.isDetectParagraphLanguage || (currentPageLanguageByClient2 = getCurrentPageLanguageByClient());
+    for (let pairKey of pariKeys) {
+      let targetPair = ctx.rule.pairs[pairKey], sourceElements = rootFrame.querySelectorAll(
+        pairKey
+      ), targetElements = rootFrame.querySelectorAll(
+        targetPair
+      );
+      for (let i3 = 0; i3 < sourceElements.length; i3++) {
+        let sourceElement = sourceElements[i3], targetElement = targetElements[i3];
+        if (sourceElement && targetElement) {
+          let paragraph = {
+            elements: [sourceElement],
+            isVertical: !1,
+            rootFrame,
+            text: sourceElement.textContent || "",
+            variables: [],
+            inline: !1,
+            preWhitespace: !0,
+            isPdf: !1,
+            targetContainer: targetElement,
+            id: id++,
+            languageByClient: currentPageLanguageByClient2,
+            languageByLocal: currentPageLanguage2
+          };
+          addToParagraphs(paragraph, allParagraphs);
+        }
+      }
+    }
+    return allParagraphs;
+  }
+
   // dom/get_pdf_paragraphs.ts
-  async function getParagraphs2(rootFrame, containers, ctx, targetContainers) {
+  async function getParagraphs3(rootFrame, containers, ctx, targetContainers) {
     let allParagraphs = [], { rule } = ctx;
     for (let i3 = 0; i3 < containers.length; i3++) {
       let container = containers[i3], targetContainer = targetContainers[i3];
@@ -14008,7 +14120,7 @@ ${injectedCss}}
     constructor() {
       super(...arguments);
       this.isSupportList = !0;
-      this.maxTextLength = 5e3;
+      this.maxTextLength = 1200;
     }
     async translate(payload) {
       let { text } = payload, startSpaceLength = text.match(/^\s*/)[0].length;
@@ -14667,7 +14779,7 @@ ${injectedCss}}
       serviceConfig.APIKEY && (this.APIKEY = serviceConfig.APIKEY?.trim(), this.apiKeys = this.APIKEY.split(",").map((key) => key.trim())), serviceConfig.prompt && (this.prompt = serviceConfig.prompt), serviceConfig.model && (this.model = serviceConfig.model), serviceConfig && serviceConfig.apiUrl && (this.apiUrl = mergeUrl(this.apiUrl, serviceConfig.apiUrl)), serviceConfig && serviceConfig.systemPrompt && (this.systemPrompt = serviceConfig.systemPrompt);
     }
     getDefaultRateLimit() {
-      return { limit: 10, interval: 65e3 };
+      return { limit: 1500, interval: 65e3 };
     }
     translate(payload) {
       return this.model.includes("003") ? (this.maxTextGroupLength = 1, this.translate3(payload)) : this.translate3_5(payload);
@@ -16290,7 +16402,7 @@ ${injectedCss}}
       translateNewDynamicNodes,
       3e3
     )
-  }, env3 = getEnv(), isProd2 = env3.PROD === "1", isInitTranslationService = !1, titleMutationObserver, mutationObserverMap = /* @__PURE__ */ new Map(), mainMutaionObserver, originalPageTitle = "";
+  }, env3 = getEnv(), isProd2 = env3.PROD === "1", translationServiceInitmap = {}, titleMutationObserver, mutationObserverMap = /* @__PURE__ */ new Map(), mainMutaionObserver, originalPageTitle = "";
   async function toggleTranslatePage() {
     if (getPageStatus() === "Original") {
       let ctx = await getGlobalContext(getRealUrl(), {});
@@ -16439,14 +16551,27 @@ ${injectedCss}}
   }
   function translationParagraph(visibleParagraph, ctx) {
     waitToTranslateParagraphIds.delete(visibleParagraph.id);
-    let paragraphWithState = getParagraph(visibleParagraph.id);
+    let pairKeys = Object.keys(ctx.rule.pairs), paragraphWithState = getParagraph(visibleParagraph.id);
     paragraphWithState && (paragraphWithState.observers && paragraphWithState.observers.length > 0 && paragraphWithState.observers.forEach((observer) => {
       observer.disconnect();
     }), paragraphWithState.observers = [], paragraphWithState.state = "Translating", setParagraph(visibleParagraph.id, paragraphWithState));
     let id = visibleParagraph.id;
     currentTranslatedTextLength += visibleParagraph.text.length;
     let realElements = getHTMLElements(visibleParagraph.elements);
-    if (visibleParagraph.isPdf) {
+    if (pairKeys.length > 0) {
+      let targetTranslationContainer = visibleParagraph.targetContainer, targetTranslationWrapper = document.createElement(
+        ctx.rule.targetWrapperTag
+      );
+      targetTranslationContainer.appendChild(targetTranslationWrapper), targetTranslationWrapper.classList.add(
+        "notranslate",
+        translationTargetElementWrapperClass
+      ), targetTranslationWrapper.id = `${translationTargetElementWrapperClass}-${id}`, targetTranslationWrapper.setAttribute("lang", ctx.targetLanguage);
+      let loadingHtml = getLoadingHTML(
+        ctx.config.loadingTheme,
+        ctx.rule
+      );
+      targetTranslationWrapper.innerHTML = loadingHtml;
+    } else if (visibleParagraph.isPdf) {
       let firstElement = getFirstHTMLElement(visibleParagraph.elements), elementStyle = globalThis.getComputedStyle(firstElement), top = elementStyle.top, fontSize = elementStyle.fontSize, fontSizeNumber = parseFloat(fontSize.slice(0, -2));
       isNaN(fontSizeNumber) || fontSizeNumber > 20 && (fontSize = "20px");
       let targetContainer = visibleParagraph.targetContainer, paragraphTarget = document.createElement(ctx.rule.targetWrapperTag);
@@ -16490,7 +16615,7 @@ ${injectedCss}}
     currentParagraphIds.push(id), throttleMap[ctx.state.translationDebounce](ctx);
   }
   function addParagraphToQueue(paragraph, ctx) {
-    ctx.state.translationStartMode === "dynamic" && !ctx.rule.isEbookBuilder && currentTranslatedTextLength > ctx.state.immediateTranslationTextCount ? onElementVisible(paragraph, (visibleParagraph) => {
+    ctx.state.translationStartMode === "dynamic" && !ctx.rule.isEbookBuilder && !ctx.rule.isSubtitleBuilder && currentTranslatedTextLength > ctx.state.immediateTranslationTextCount ? onElementVisible(paragraph, (visibleParagraph) => {
       ctx.rule.visibleDelay > 0 ? setTimeout(() => {
         translationParagraph(visibleParagraph, ctx);
       }, ctx.rule.visibleDelay) : translationParagraph(visibleParagraph, ctx);
@@ -16509,6 +16634,15 @@ ${injectedCss}}
         detail: JSON.stringify({
           type: "translateStart",
           payload: {}
+        })
+      })
+    ), document.dispatchEvent(
+      new CustomEvent(documentMessageTypeIdentifierForTellThirdParty, {
+        detail: JSON.stringify({
+          type: "targetLanguage",
+          payload: {
+            targetLanguage: ctx.targetLanguage
+          }
         })
       })
     ), ctx.rule.normalizeBody && document.querySelector(ctx.rule.normalizeBody)) {
@@ -16533,9 +16667,25 @@ ${injectedCss}}
     }
     ctx.state.isAutoTranslate = !0;
     let currentScrollOffset = globalThis.scrollY, currentWindowHeight = globalThis.innerHeight;
-    currentScrollOffset >= currentWindowHeight && (ctx.state.immediateTranslationTextCount = 0), isInitTranslationService || (isInitTranslationService = !0, isInIframe || initTranslationEngine(ctx).catch((e) => {
+    if (currentScrollOffset >= currentWindowHeight && (ctx.state.immediateTranslationTextCount = 0), translationServiceInitmap[ctx.translationService] || (translationServiceInitmap[ctx.translationService] = !0, isInIframe || initTranslationEngine(ctx).catch((e) => {
       log_default.warn("init translation engine error", e);
-    })), log_default.debug("ctx", ctx), addToUnmountQueue(() => {
+    })), log_default.debug("ctx", ctx), Object.keys(ctx.rule.pairs).length > 0) {
+      let paragraphs = getParagraphs2(ctx.mainFrame, ctx);
+      document.dispatchEvent(
+        new CustomEvent(documentMessageTypeIdentifierForTellThirdParty, {
+          detail: JSON.stringify({
+            type: "totalParagraphsCount",
+            payload: {
+              totalParagraphsCount: paragraphs.length
+            }
+          })
+        })
+      );
+      for (let paragraph of paragraphs)
+        addParagraphToQueue(paragraph, ctx);
+      return;
+    }
+    addToUnmountQueue(() => {
       currentTranslatedTextLength = 0, cleanParagraphs(), allIntersectionObserver.forEach((observer) => {
         observer.disconnect();
       }), allResizebleObserver.forEach((observer) => {
@@ -16693,7 +16843,7 @@ ${injectedCss}}
       setPageTranslatedStatus("Translating");
     }
     let paragraphs = [];
-    if (ctx.rule.isPdf ? paragraphs = await getParagraphs2(
+    if (ctx.rule.isPdf ? paragraphs = await getParagraphs3(
       rootFrame,
       containers,
       ctx,
@@ -16712,6 +16862,14 @@ ${injectedCss}}
     setPageTranslatedStatus("Translated");
   }
   async function retryFailedParagraphs() {
+    document.dispatchEvent(
+      new CustomEvent(documentMessageTypeIdentifierForTellThirdParty, {
+        detail: JSON.stringify({
+          type: "retryFailedParagraphsStart",
+          payload: {}
+        })
+      })
+    );
     let allParagraphEntities = getParagraphEntities(), ids = [];
     for (let [id, paragraph] of allParagraphEntities)
       paragraph.state === "Error" && ids.push(id);
@@ -16960,7 +17118,7 @@ ${injectedCss}}
     else if (log_default.debug("do not auto translate", ctx), ctx.rule.initTranslationServiceAsSoonAsPossible && ctx.translationService === "deepl") {
       if (isSameTargetLanguage(lang, ctx.targetLanguage) || lang === "auto")
         return;
-      ctx.config && ctx.config.translationServices && ctx.config.translationServices.deepl && ctx.config.translationServices.deepl.authKey && typeof ctx.config.translationServices.deepl.authKey == "string" && ctx.config.translationServices.deepl.authKey.startsWith("immersive_") && (isInitTranslationService || (isInitTranslationService = !0, isInIframe || initTranslationEngine(ctx).catch((e) => {
+      ctx.config && ctx.config.translationServices && ctx.config.translationServices.deepl && ctx.config.translationServices.deepl.authKey && typeof ctx.config.translationServices.deepl.authKey == "string" && ctx.config.translationServices.deepl.authKey.startsWith("immersive_") && (translationServiceInitmap[ctx.translationService] || (translationServiceInitmap[ctx.translationService] = !0, isInIframe || initTranslationEngine(ctx).catch((e) => {
         log_default.warn("init translation engine error", e);
       })));
     }
@@ -16987,10 +17145,10 @@ ${injectedCss}}
   }
   async function detectCurrentPageLanguage(ctx) {
     ctx || (globalContext = await getGlobalContext(getRealUrl(), {}), ctx = globalContext);
-    let isInIframe = getIsInIframe(), lang = "auto";
+    let isInIframe = getIsInIframe(), lang = "auto", pairKeys = Object.keys(ctx.rule.pairs);
     if (isMonkey()) {
       let mainText = "";
-      ctx.rule.isEbook || ctx.rule.isEbookBuilder ? mainText = getAllIframeMainText(ctx.mainFrame) : mainText = getMainText(ctx.mainFrame).slice(0, 1e3), lang = await detectLanguage({
+      ctx.rule.isEbook || ctx.rule.isEbookBuilder ? mainText = getAllIframeMainText(ctx.mainFrame) : pairKeys.length > 0 ? mainText = getPairsSourceText(ctx.mainFrame, ctx.rule.pairs) : mainText = getMainText(ctx.mainFrame).slice(0, 1e3), lang = await detectLanguage({
         text: mainText
       });
     } else if (isInIframe)
@@ -17000,6 +17158,11 @@ ${injectedCss}}
     else if (ctx.rule.isEbook || ctx.rule.isEbookBuilder) {
       let mainText = "";
       mainText = getAllIframeMainText(ctx.mainFrame), lang = await detectLanguage({
+        text: mainText
+      });
+    } else if (pairKeys.length > 0) {
+      let mainText = getPairsSourceText(ctx.mainFrame, ctx.rule.pairs);
+      lang = await detectLanguage({
         text: mainText
       });
     } else
@@ -17962,6 +18125,7 @@ ${injectedCss}}
       onUserConfigChange,
       request: request3,
       onTranslateLocalHtmlFile,
+      onTranslateLocalSubtitleFile,
       onSetBuildinConfig,
       pageStatus: pageStatus2,
       openEbookBuilderPage: openEbookBuilderPage3,
@@ -18407,7 +18571,7 @@ ${injectedCss}}
             label: t4("more"),
             showArrow: !0,
             onSelected: (item) => {
-              item.value === "translateTheWholePage" ? onTranslateTheWholePage() : item.value === "translateToThePageEndImmediately" ? ontranslateToThePageEndImmediately() : item.value === "translateTheMainPage" ? onTranslateTheMainPage() : item.value === "showTranslationOnly" || (item.value === "translateLocalPdfFile" ? onTranslateLocalPdfFile && onTranslateLocalPdfFile() : item.value === "translateLocalHtmlFile" ? onTranslateLocalHtmlFile && onTranslateLocalHtmlFile() : item.value === "donate" ? (globalThis.open(config.donateUrl), onClose()) : item.value === "feedback" ? (globalThis.open(config.feedbackUrl), onClose()) : item.value === "options" ? (openOptionsPage3(), onClose()) : item.value === "changeToTranslateTheWholePage" ? handleChangeToTranslateTheWholePage() : item.value === "changeToTranslateTheMainPage" ? handleChangeToTranslateTheMainPage() : item.value === "about" ? openAboutPage3() : item.value === "toggleEnabled" ? onToggleEnabled() : item.value === "openEbookViewer" ? openEbookViewerPage3() : item.value === "openEbookBuilder" && openEbookBuilderPage3());
+              item.value === "translateTheWholePage" ? onTranslateTheWholePage() : item.value === "translateToThePageEndImmediately" ? ontranslateToThePageEndImmediately() : item.value === "translateTheMainPage" ? onTranslateTheMainPage() : item.value === "showTranslationOnly" || (item.value === "translateLocalPdfFile" ? onTranslateLocalPdfFile && onTranslateLocalPdfFile() : item.value === "translateLocalHtmlFile" ? onTranslateLocalHtmlFile && onTranslateLocalHtmlFile() : item.value === "translateLocalSubtitleFile" ? onTranslateLocalSubtitleFile && onTranslateLocalSubtitleFile() : item.value === "donate" ? (globalThis.open(config.donateUrl), onClose()) : item.value === "feedback" ? (globalThis.open(config.feedbackUrl), onClose()) : item.value === "options" ? (openOptionsPage3(), onClose()) : item.value === "changeToTranslateTheWholePage" ? handleChangeToTranslateTheWholePage() : item.value === "changeToTranslateTheMainPage" ? handleChangeToTranslateTheMainPage() : item.value === "about" ? openAboutPage3() : item.value === "toggleEnabled" ? onToggleEnabled() : item.value === "openEbookViewer" ? openEbookViewerPage3() : item.value === "openEbookBuilder" && openEbookBuilderPage3());
             },
             menus: [
               config.translationArea === "main" && {
@@ -18437,6 +18601,10 @@ ${injectedCss}}
               !isMonkey() && {
                 label: "\u{1F310} " + t4("browser.translateLocalHtmlFile"),
                 value: "translateLocalHtmlFile"
+              },
+              {
+                label: "\u{1F4FA} " + t4("browser.translateLocalSubtitleFile"),
+                value: "translateLocalSubtitleFile"
               },
               {
                 label: (config.enabled ? "\u{1F6AB} " : "\u{1F44B} ") + (config.enabled ? t4("clickToDisableExtension") : t4("clickToEnableExtension")),
@@ -18660,6 +18828,10 @@ ${injectedCss}}
       openEbookBuilderPage2(), setTimeout(() => {
         onClose();
       }, 50);
+    }, handleOpenSubtitleBuilderPage = () => {
+      openSubtitleBuilderPage2(), setTimeout(() => {
+        onClose();
+      }, 50);
     };
     return !config || !ctx ? null : /* @__PURE__ */ p5(
       Popup,
@@ -18668,6 +18840,7 @@ ${injectedCss}}
         openEbookBuilderPage: handleOpenEbookBuilderPage,
         onTranslateLocalHtmlFile: () => {
         },
+        onTranslateLocalSubtitleFile: handleOpenSubtitleBuilderPage,
         request: request2,
         onClose: handleClose,
         onToggleEnabled: handleToggleEnabled,
@@ -18931,7 +19104,7 @@ ${injectedCss}}
           });
         });
       } else
-        ctx.rule.isEbook || ctx.rule.isEbookBuilder || await initPage();
+        ctx.rule.isEbook || ctx.rule.isEbookBuilder || ctx.rule.isSubtitleBuilder || await initPage();
       checkCronAndRunOnce();
     }
   }
@@ -19149,7 +19322,7 @@ ${this._lastError.message}`;
     manifest_version: 3,
     name: "__MSG_brandName__",
     description: "__MSG_brandDescription__",
-    version: "0.3.15",
+    version: "0.4.0",
     default_locale: "en",
     background: {
       service_worker: "background.js"