+++ scripts/sponsorkit/src/start.ts
@@ -136,6 +136,7 @@ async function fetchCustomSponsors(filepath: string) {
       const parts = trimed.split(",");
       let name = "";
       let avatar = "";
+      let profile = "";
       let expired = 0;
       let monthlyDollars = 0;
       let startDate = 0;
@@ -165,6 +166,13 @@ async function fetchCustomSponsors(filepath: string) {
             case 3:
               startDate = expired -
                 1000 * 60 * 60 * 24 * 30 * parseInt(trimedPart);
+              break;
+            case 4:
+              avatar = trimedPart;
+              break;
+            case 5:
+              profile = trimedPart;
+              break;
           }
         }
       }
@@ -174,27 +182,31 @@ async function fetchCustomSponsors(filepath: string) {
       }
 
       const isExpired = expired && expired < Date.now();
-
-      // add to sponsors
-      const sponsor: Sponsorship = {
-        sponsor: {
-          type: "User",
-          login: "",
-          name: name,
-          avatarUrl: "",
-          avatarUrlHighRes: "",
-          avatarUrlMediumRes: "",
-          avatarUrlLowRes: "",
-          websiteUrl: "",
-          linkUrl: "",
-        },
-        monthlyDollars: isExpired ? -1 : monthlyDollars,
-        privacyLevel: "PUBLIC",
-        createdAt: new Date(startDate).toISOString(),
-        isOneTime: true,
-        provider: "custom",
-      };
-      sponsors.push(sponsor);
+      try {
+        // add to sponsors
+        const sponsor: Sponsorship = {
+          sponsor: {
+            type: "User",
+            login: "",
+            name: name,
+            avatarUrl: avatar,
+            avatarUrlHighRes: "",
+            avatarUrlMediumRes: "",
+            avatarUrlLowRes: "",
+            websiteUrl: "",
+            linkUrl: profile,
+          },
+          monthlyDollars: isExpired ? -1 : monthlyDollars,
+          privacyLevel: "PUBLIC",
+          createdAt: new Date(startDate).toISOString(),
+          isOneTime: true,
+          provider: "custom",
+        };
+        sponsors.push(sponsor);
+      } catch (e) {
+        console.error("startDate", filepath, startDate, name);
+        throw e;
+      }
     }
   }